[Win32] Move all monitor-specific scaling functionality to DPIUtil
DPIUtil already contains all monitor-specific-scaling-related
functionality for retrieving the state and identifying compatibility.
Only the (internal) possibility to set the mode is placed in
Win32DPIUtils, which just scatters the according logic.

This change moves the monitor-specific-scaling-related utility functions
completely to DPIUtil. On Linux and MacOS they will currently not have
an effect.
