+++ bundles/org.eclipse.swt/Eclipse SWT/win32/org/eclipse/swt/widgets/Display.java
@@ -951,7 +951,7 @@ protected void create (DeviceData data) {
 	checkDisplay (thread = Thread.currentThread (), true);
 	if (DPIUtil.isMonitorSpecificScalingActive()) {
 		setMonitorSpecificScaling(true);
-		Win32DPIUtils.setAutoScaleForMonitorSpecificScaling();
+		DPIUtil.setMonitorSpecificScaling(true);
 	}
 	Win32DPIUtils.initializeCustomDpiAwareness();
 	createDisplay (data);
